Reverting changes related to number types. Better approach envisioned, does not require so many changes

+++ src/action/fixtures/simple_action_df.sql
@@ -27,7 +27,7 @@ SET default_with_oids = false;
 CREATE TABLE "__ONTASK_WORKFLOW_TABLE_5" (
     age double precision,
     email text,
-    sid double precision,
+    sid bigint,
     another text,
     name text,
     one text,
